Added development dependencies and uml directory
This commit is barely complete, there is an awful lot
I still ignore about hybrid app testing and debugging.

These tools are not definitive and might be excluded/replaced
in the future.

Tools added in this commit:
  > calabash-android.

About the uml directory, the uml diagrams are created with
the staruml tool.