Added codebase for scraper module
This commit contains code needed by the scraper module of the product.
I made some changes to the UML diagrams as I get acquainted of how
javascript workd, since it's my first project.

This commit is not final, as I get acquainted of how javascript works
and what the best practices are, I might end up changing most of the
code included in this commit.