Refactoring and additions to support scraping.
The modules that are under development needed the unit tests to asure
the code worked as expected. The need for a code style was obvious since
the start, but getting to know AngularJS helped choose the most suitable
code style.

This commit includes the unit tests for the wizzle.core and
wizzle.scraper modules, which explains the addition of the karma.conf.js
file. Also some refactoring has been done to wizzle.core,
wizzle.scraper and wizzle.utils. These changes are in the intention of
adopting a coding style (google code style).